Move other relevant hotbar classes.
